fix: NextJS から使用できるように修正
